# openapi.yaml
openapi: 3.0.3
info:
  title: YieldScan Backend API
  description: |
    API for managing crypto assets, their yield-generating steps, and retrieving APY data.
    This API supports protocol-based step organization with comprehensive asset and step management capabilities.
  version: 1.0.0
  contact:
    name: YieldScan Team
servers:
  - url: http://localhost:4023/api
    description: Local development server
  - url: https://api.yieldscan.com/api
    description: Production server
paths:
...
  /assets/{id}:
    get:
      tags:
        - Assets
      summary: Get asset by ID
      description: Retrieve a specific asset by its ID
      parameters:
        - name: id
          in: path
          required: true
          description: Asset ID
          schema:
            type: string
            example: 684a8cf6db44b98888f22f5f
        - name: includeDisabled
          in: query
          description: Include disabled assets in results
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  asset:
                    $ref: '#/components/schemas/Asset'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    put:
      tags:
        - Assets
      summary: Update asset
      description: Update an existing asset
      parameters:
        - name: id
          in: path
          required: true
          description: Asset ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAssetRequest'
      responses:
        '200':
          description: Asset updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Asset updated successfully
                  asset:
                    $ref: '#/components/schemas/Asset'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    delete:
      tags:
        - Assets
      summary: Delete asset
      description: Delete an asset by ID
      parameters:
        - name: id
          in: path
          required: true
          description: Asset ID
          schema:
            type: string
      responses:
        '200':
          description: Asset deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Asset deleted successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
    Asset:
      type: object
      properties:
        _id:
          type: string
          example: 684a8cf6db44b98888f22f5f
        token:
          type: string
          example: USDT
        chain:
          type: string
          example: ETH
        address:
          type: string
          example: "0xdAC17F958D2ee523a2206206994597C13D831ec7"
        icon:
          type: string
          example: "https://cdn.jsdelivr.net/gh/trustwallet/assets/blockchains/ethereum/assets/0xdAC17F958D2ee523a2206206994597C13D831ec7/logo.png"
        chainId:
          type: integer
          example: 1
        decimals:
          type: integer
          example: 6
        maxDecimalsShow:
          type: integer
          example: 2
        usdPrice:
          type: number
          example: 1.0
        yieldBearingToken:
          type: boolean
          example: false
        enabled:
          type: boolean
          example: true
        underlyingAsset:
          type: string
          nullable: true
          example: null
        withdrawContract:
          type: string
          nullable: true
          example: null
        protocol:
          type: string
          nullable: true
          example: null
        withdrawUri:
          type: string
          nullable: true
          example: null
        maturity:
          type: string
          nullable: true
          example: null
        lockYield:
          $ref: '#/components/schemas/LockYield'
        createdAt:
          type: string
          format: date-time
          example: "2025-06-16T08:11:29.443Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-06-16T08:25:35.229Z"
...
    UpdateAssetRequest:
      type: object
      properties:
        token:
          type: string
        chain:
          type: string
        address:
          type: string
        icon:
          type: string
        chainId:
          type: integer
        decimals:
          type: integer
        maxDecimalsShow:
          type: integer
        usdPrice:
          type: number
        yieldBearingToken:
          type: boolean
        enabled:
          type: boolean
        underlyingAsset:
          type: string
          nullable: true
        withdrawContract:
          type: string
          nullable: true
        protocol:
          type: string
          nullable: true
        withdrawUri:
          type: string
          nullable: true
        maturity:
          type: string
          nullable: true
        lockYield:
          $ref: '#/components/schemas/LockYield'
...
    LockYield:
      type: object
      nullable: true
      properties:
        expirationDate:
          type: string
          example: "2024-12-31"
        protocol:
          $ref: '#/components/schemas/LockYieldProtocol'
    LockYieldProtocol:
      type: object
      properties:
        name:
          type: string
          example: "Pendle"
        swap:
          type: boolean
          example: true
        ytAddress:
          type: string
          example: "0x..."
        ptAddress:
          type: string
          example: "0x..."
        swapAddress:
          type: string
          example: "0x..."
        ytMarketAddress:
          type: string
          example: "0x..."
        ytDecimals:
          type: integer
          example: 18
        ptDecimals:
          type: integer
          example: 18
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Validation failed"
        details:
          type: string
          example: "Missing required field: token"
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Missing required fields"
            details: "assetId, order, title, description, chain, functionConfig are required"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Asset not found"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Failed to fetch assets"
            details: "Database connection error"
...